requirements.txt
openai==0.27.6
curl_cffi
bs4
requests
sqlalchemy
python-dotenv==1.0.0
pygsheets
pandas
lxml
google-generativeai
grpcio
grpcio-tools
selenium

gspread
pandas
gspread_dataframe
google-auth
google-auth-oauthlib
google-auth-httplib2

langchain
langchain-community
langchain-core
langchain-google-genai
google-generativeai
jinja2

Django>=4.2,<5.0
psycopg2-binary>=2.9,<3.0

django-crontab
